LayoutTests/fast/parser/block-nesting-cap-expected.txt

Test that the HTML parser does not allow the nesting depth of "block-level" elements to exceed 4096.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS d4096.parentNode === d4094 is true
PASS d4095.parentNode === d4094 is true
PASS d4096.previousSibling === d4095 is true
PASS successfullyParsed is true

TEST COMPLETE
